+++ adminSiteServer/webpackUtils.tsx
@@ -5,7 +5,7 @@ import urljoin from "url-join"
 import * as path from "path"
 
 let manifest: { [key: string]: string }
-export const getWebpackUrlForAsset = (assetName: string) => {
+export const webpackUrl = (assetName: string) => {
     if (ENV === "production") {
         // Read the real asset name from the manifest in case it has a hashed filename
         if (!manifest)
@@ -27,3 +27,32 @@ export const getWebpackUrlForAsset = (assetName: string) => {
 
     return urljoin(WEBPACK_DEV_URL, assetName)
 }
+
+export const bakeEmbedSnippet = () => `const embedSnippet = () => {
+const link = document.createElement('link')
+link.type = 'text/css'
+link.rel = 'stylesheet'
+link.href = '${webpackUrl("commons.css")}'
+document.head.appendChild(link)
+
+let loadedScripts = 0;
+const checkReady = () => {
+    loadedScripts++
+    if (loadedScripts === 3)
+        window.MultiEmbedderSingleton.embedAll()
+}
+
+const coreScripts = ['https://cdn.polyfill.io/v2/polyfill.min.js?features=es6,fetch', '${webpackUrl(
+    "commons.js"
+)}', '${webpackUrl("owid.js")}']
+
+coreScripts.forEach(url => {
+    const script = document.createElement('script')
+    script.type = 'text/javascript'
+    script.onload = checkReady
+    script.src = url
+    document.head.appendChild(script)
+})
+}
+embedSnippet()
+`
