+++ baker/SiteBaker.tsx
@@ -34,7 +34,6 @@ import {
 } from "site/grapherUtil"
 import { makeSitemap } from "site/sitemap"
 import * as React from "react"
-import { bakeEmbedSnippet } from "./bakeEmbedSnippet"
 import { Post } from "db/model/Post"
 import { bakeCountries } from "site/countryProfiles"
 import { countries } from "clientUtils/countries"
@@ -52,7 +51,7 @@ import {
 import { getRedirects } from "./redirects"
 import { bakeAllChangedGrapherPagesVariablesPngSvgAndDeleteRemovedGraphers } from "./GrapherBaker"
 import { EXPLORERS_ROUTE_FOLDER } from "explorer/ExplorerConstants"
-import { getWebpackUrlForAsset } from "adminSiteServer/staticGen"
+import { bakeEmbedSnippet } from "adminSiteServer/webpackUtils"
 
 export class SiteBaker {
     private grapherExports!: GrapherExports
@@ -305,9 +304,7 @@ export class SiteBaker {
 
         await fs.writeFile(
             `${this.bakedSiteDir}/grapher/embedCharts.js`,
-            bakeEmbedSnippet(
-                getWebpackUrlForAsset("commons.css", "commons.js", "owid.js")
-            )
+            bakeEmbedSnippet()
         )
         this.stage(`${this.bakedSiteDir}/grapher/embedCharts.js`)
     }
